fix(mcp): move get_next_recommended_work to the read-tools module
Two merge-blocking guards caught real problems in the previous commit.

Module Size Guard: backlog-pack.ts was 798 LOC on main against a hard 800
ceiling, so the caveat wiring pushed it to 805. Rather than shave the change to
squeak under — which would leave the next contributor at the same wall —
get_next_recommended_work moves to backlog-pack-read-tools.ts, where its four
sibling read handlers (queryBacklog, listEpics, listBacklogItems,
getBacklogItem) already live and where backlog-pack.ts already imports them
from. Pure move plus import wiring; no behaviour change. 805 -> 678 and
489 -> 622, both well clear of the ceiling.

Doc Anchor Existence: the runbook cited BI-72F368BC, which does not exist in
the live backlog. The citation is pre-existing, not introduced here — touching
the file is what brought it under the gate. The rule it annotates is
independently cited to the initiative-readiness design spec, so the dead id is
dropped rather than the rule.

Verification: 172 test files / 1576 tests still pass; `pnpm --filter web build`
compiles successfully (146/146 pages); module-size, doc-anchor and the 37-guard
repo loop all pass locally.
