+++ docs/architecture/backlog-and-planning-runbook.md
@@ -9,7 +9,7 @@
 - **Capture is the other half of recovery, and it goes stale.** Every epic with a not-done item must have a *current* bundle in [`packages/db/recovery/backlog/`](../../packages/db/recovery/backlog/README.md), merged to `main` — a bundle on an unmerged branch is not a backup, and a stale bundle is worse than a missing one because the epic still looks protected while the items filed since capture are exactly the ones lost. Re-capture with `pnpm --filter @dpf/db backlog:capture -- --out packages/db/recovery/backlog` immediately before any reset or teardown, and check `manifest.json` closes: `capturedItemCount + unassignedItemCount == unfinishedItemCount`. Items with no epic cannot be bundled (the format is epic-scoped and requires `epicId`) — link them with `link_backlog_item_to_epic` first, or accept they survive only as the non-reconcilable `unassigned-items.json` record.
 - **Specs and plans** live in `docs/superpowers/specs/` and `docs/superpowers/plans/`. Check for an existing design before starting work — some are ready to implement.
 - **Plans need live backlog coverage before implementation.** Use `dpf-writing-plans` and the governed `record_plan_backlog_coverage` receipt. Every independently shippable deliverable maps to a live new or existing `BacklogItem`; an `xlarge` BI always records a decomposition decision. Keeping one BI is allowed only with an auditable rationale that its phases are sequencing, not independent work. Copy the receipt, BI mappings, and dependencies into the plan. MCP/tool/scope failure stops the workflow — Markdown checkboxes are never backlog coverage. The plugin pre-source guard is the earliest hard boundary; the Plan Backlog Coverage CI check is the repository backstop.
-- **A schema-v2 coverage receipt needs a scope baseline, and only an independent reviewer can mint one (BI-72F368BC).** `record_plan_backlog_coverage` binds the plan to the umbrella item's `initiative_scope_baseline`. A baseline exists only after the initiative's spec-approval gate passes: `record_initiative_design_review` with `gate: "spec-approval"`, `decision: "pass"`, and a canonical design under `docs/superpowers/specs/` whose objectives are marked `**OBJ-*:**` and whose acceptance statements are `| AC-* | OBJ-… | statement |` table rows (the server parses the manifest; the caller never supplies it). That lane requires a reviewer **independent of the artifact author**, and independence binds between *principals*: an author cannot approve their own design, superuser status does not waive it, and there is deliberately no break-glass self-approval path — adding one needs its own design with second approval, expiry, rationale and durable audit ([initiative-readiness design](../superpowers/specs/2026-08-08-initiative-readiness-and-goal-completion-reconciliation-design.md) §361).
+- **A schema-v2 coverage receipt needs a scope baseline, and only an independent reviewer can mint one.** `record_plan_backlog_coverage` binds the plan to the umbrella item's `initiative_scope_baseline`. A baseline exists only after the initiative's spec-approval gate passes: `record_initiative_design_review` with `gate: "spec-approval"`, `decision: "pass"`, and a canonical design under `docs/superpowers/specs/` whose objectives are marked `**OBJ-*:**` and whose acceptance statements are `| AC-* | OBJ-… | statement |` table rows (the server parses the manifest; the caller never supplies it). That lane requires a reviewer **independent of the artifact author**, and independence binds between *principals*: an author cannot approve their own design, superuser status does not waive it, and there is deliberately no break-glass self-approval path — adding one needs its own design with second approval, expiry, rationale and durable audit ([initiative-readiness design](../superpowers/specs/2026-08-08-initiative-readiness-and-goal-completion-reconciliation-design.md) §361).
 
   **On an install with one human principal, the independent reviewer is an in-platform reviewer coworker, not another person.** A coworker has its own `Principal` (`kind: agent`) and its own `initiative_design_review` grant; summon it and have *it* record the gate, and the receipt is attributed to the coworker's principal rather than to the delegating human. Recording the same call yourself is a self-review and is refused — correctly. This is the paved road, not a workaround: the review is by a differently-grounded reviewer with its own accountable identity and its own audit trail. Before the fix, every reviewer resolved to the human regardless of who called, so no install could produce a baseline and no v2 coverage receipt existed anywhere.
 - **Enforced process/docs/data CI gates (mandatory, surface-agnostic).** A substantial implementation surface needs a spec/plan/doc touch or a `Process-Spine-Decision:` trailer; a UX / workflow / queue / nav / attention surface needs a `## Design grounding` section or `Design-Grounding-Decision:` trailer naming the specs/plans AND code substrate reviewed; a documented user-facing route change needs the doc edit or a `Docs-Impact-Decision:` trailer; a persistent Prisma model / migration / projection / AI-context source needs a generated `*.data-impact.json` manifest (or exception) AND a `packages/db/src/table-classification.ts` sensitivity + exactly one retention disposition. Internal doc links are source-relative `.md` paths — regenerate with `pnpm docs:index` on add/rename/delete. **Each gate self-enforces in CI, so a missing trailer fails the PR, not your memory** — the reason this detail need not stay always-on. Full per-gate trigger + trailer mechanics (Spec/Plan/Doc, Design Grounding, Doc Reference Integrity, Docs Link, Docs Impact, Data-Impact, Stewardship Scope): [`docs/architecture/enforced-ci-gates.md`](../architecture/enforced-ci-gates.md).
